refactor: move multi-tenant job worker element tests to separate class
The job worker element-related multi-tenant test cases were written
when the engine unit tests were tenant-unaware. As a
result, it was possible to work with custom tenants
without enabling multi-tenancy. With the recent
refactoring of the engine to be tenant-aware, we
need to explicitly enable multi-tenancy in the test
classes for the custom tenant test cases to work
correctly. These tests are now moved in a dedicated
multi-tenant test class.
