+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/bpmn/activity/JobWorkerElementTest.java
@@ -23,12 +23,10 @@ import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.protocol.record.intent.VariableIntent;
 import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import io.camunda.zeebe.protocol.record.value.EntityType;
 import io.camunda.zeebe.protocol.record.value.JobKind;
 import io.camunda.zeebe.protocol.record.value.JobListenerEventType;
 import io.camunda.zeebe.protocol.record.value.JobRecordValue;
 import io.camunda.zeebe.protocol.record.value.ProcessInstanceRecordValue;
-import io.camunda.zeebe.test.util.Strings;
 import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
 import java.util.Collection;
@@ -46,6 +44,8 @@ import org.junit.runners.Parameterized.Parameters;
 /**
  * Verifies the behavior of elements that are based on jobs and should be processed by job workers.
  * For example, service tasks.
+ *
+ * <p>See {@link JobWorkerElementMultiTenantTest} for the non-multi-tenant tests.
  */
 @RunWith(Parameterized.class)
 public final class JobWorkerElementTest {
@@ -106,38 +106,6 @@ public final class JobWorkerElementTest {
         .hasFlowScopeKey(processInstanceKey)
         .hasBpmnProcessId("process")
         .hasProcessInstanceKey(processInstanceKey);
-  }
-
-  @Test
-  public void shouldActivateTaskWithCustomTenant() {
-    // given
-    final String tenantId = "foo";
-    ENGINE
-        .deployment()
-        .withXmlResource(process(t -> t.zeebeJobType("test")))
-        .withTenantId(tenantId)
-        .deploy();
-
-    // when
-    final long processInstanceKey =
-        ENGINE.processInstance().ofBpmnProcessId(PROCESS_ID).withTenantId(tenantId).create();
-
-    // then
-    final Record<ProcessInstanceRecordValue> taskActivating =
-        RecordingExporter.processInstanceRecords()
-            .withProcessInstanceKey(processInstanceKey)
-            .withTenantId(tenantId)
-            .withIntent(ProcessInstanceIntent.ELEMENT_ACTIVATING)
-            .withElementType(elementBuilder.getElementType())
-            .getFirst();
-
-    Assertions.assertThat(taskActivating.getValue())
-        .hasElementId("task")
-        .hasBpmnElementType(elementBuilder.getElementType())
-        .hasFlowScopeKey(processInstanceKey)
-        .hasBpmnProcessId("process")
-        .hasProcessInstanceKey(processInstanceKey)
-        .hasTenantId(tenantId);
   }
 
   @Test
@@ -248,49 +216,6 @@ public final class JobWorkerElementTest {
     assertThat(
             RecordingExporter.processInstanceRecords()
                 .withProcessInstanceKey(processInstanceKey)
-                .limitToProcessInstanceCompleted())
-        .extracting(r -> r.getValue().getBpmnElementType(), Record::getIntent)
-        .containsSubsequence(
-            tuple(elementBuilder.getElementType(), ProcessInstanceIntent.ELEMENT_COMPLETING),
-            tuple(elementBuilder.getElementType(), ProcessInstanceIntent.ELEMENT_COMPLETED),
-            tuple(BpmnElementType.PROCESS, ProcessInstanceIntent.ELEMENT_COMPLETED));
-  }
-
-  @Test
-  public void shouldCompleteTaskWithCustomTenant() {
-    // given
-    final String tenantId = Strings.newRandomValidIdentityId();
-    final String username = Strings.newRandomValidIdentityId();
-    ENGINE.tenant().newTenant().withTenantId(tenantId).create();
-    ENGINE.user().newUser(username).create();
-    ENGINE
-        .tenant()
-        .addEntity(tenantId)
-        .withEntityId(username)
-        .withEntityType(EntityType.USER)
-        .add();
-    ENGINE
-        .deployment()
-        .withXmlResource(process(t -> t.zeebeJobType("test")))
-        .withTenantId(tenantId)
-        .deploy();
-
-    // when
-    final long processInstanceKey =
-        ENGINE.processInstance().ofBpmnProcessId(PROCESS_ID).withTenantId(tenantId).create();
-
-    ENGINE
-        .job()
-        .ofInstance(processInstanceKey)
-        .withType("test")
-        .withAuthorizedTenantIds(tenantId)
-        .complete(username);
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords()
-                .withProcessInstanceKey(processInstanceKey)
-                .withTenantId(tenantId)
                 .limitToProcessInstanceCompleted())
         .extracting(r -> r.getValue().getBpmnElementType(), Record::getIntent)
         .containsSubsequence(
